requirements.txt
streamlit==1.32.0
numpy
pillow
tflite-runtime